add fly command

+++ src/main/java/de/polo/voidroleplay/listeners/PlayerLoginListener.java
@@ -23,18 +23,23 @@ public class PlayerLoginListener implements Listener {
             Statement statement = Main.getInstance().mySQL.getStatement();
             ResultSet res = statement.executeQuery("SELECT *, DATE_FORMAT(date, '%d.%m.%Y | %H:%i:%s') AS formatted_timestamp FROM player_bans WHERE uuid = '" + player.getUniqueId() + "'");
             if (res.next()) {
+                boolean isPerm = res.getBoolean("isPermanent");
                 java.util.Date utilDate = new java.util.Date(res.getDate(6).getTime());
 
                 LocalDateTime localDateTime = utilDate.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime();
 
                 LocalDateTime date = localDateTime.atZone(ZoneId.systemDefault()).toLocalDateTime();
-                if (date.isBefore(LocalDateTime.now())) {
+                if (date.isBefore(LocalDateTime.now()) && !isPerm) {
                     event.setResult(PlayerLoginEvent.Result.ALLOWED);
                     statement.execute("DELETE FROM player_bans WHERE uuid = '" + player.getUniqueId() + "'");
                     return;
                 }
                 event.setResult(PlayerLoginEvent.Result.KICK_BANNED);
-                event.setKickMessage("§8 • §6VoidRoleplay§8 •\n\n§cDu bist vom Server gesperrt.\nGrund: " + res.getString(4) + "\nGebannt durch: " + res.getString(5) + "\nBan läuft ab: " + res.getString("formatted_timestamp")+ "\n\n§8 • §6VoidRoleplay§8 •");
+                if (!isPerm) {
+                    event.setKickMessage("§8 • §6VoidRoleplay§8 •\n\n§cDu bist vom Server gesperrt.\nGrund: " + res.getString(4) + "\nGebannt durch: " + res.getString(5) + "\nBan läuft ab: " + res.getString("formatted_timestamp")+ "\n\n§8 • §6VoidRoleplay§8 •");
+                } else {
+                    event.setKickMessage("§8 • §6VoidRoleplay§8 •\n\n§cDu bist Permanent vom Server gesperrt.\nGrund: " + res.getString(4) + "\nGebannt durch: " + res.getString(5) + "\n\n§8 • §6VoidRoleplay§8 •");
+                }
             }
         } catch (SQLException e) {
             throw new RuntimeException(e);
